# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PYCS API
  version: 0.1.0
  description: Starter OpenAPI contract for the Laravel API. Keep this synchronized with route and resource changes.
servers:
  - url: http://localhost:8000/api
paths:
  /announcements:
    get:
      summary: List announcements
      responses:
        "200":
          description: Announcement list
    post:
      summary: Create an announcement
      security:
        - sanctum: []
      responses:
        "201":
          description: Announcement created
        "401":
          description: Unauthenticated
        "422":
          description: Validation error
...
  /announcements/{announcement}:
    put:
      summary: Update an announcement
      security:
        - sanctum: []
      parameters:
        - $ref: "#/components/parameters/AnnouncementId"
      responses:
        "200":
          description: Announcement updated
        "404":
          description: Announcement not found
    delete:
      summary: Delete an announcement
      security:
        - sanctum: []
      parameters:
        - $ref: "#/components/parameters/AnnouncementId"
      responses:
        "200":
          description: Announcement deleted
        "404":
          description: Announcement not found
...
components:
  securitySchemes:
    sanctum:
      type: apiKey
      in: cookie
      name: laravel_session
  parameters:
    AnnouncementId:
      name: announcement
      in: path
      required: true
      schema:
        type: integer
...